requirements.txt
absl-py==2.3.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudpickle==3.1.2
colorama==0.4.6
dm-env==1.6
dm-tree==0.1.9
dm_control==1.0.34
etils==1.13.0
Farama-Notifications==0.0.4
filelock==3.19.1
fsspec==2025.10.0
glfw==2.10.0
gymnasium==1.2.2
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
labmaze==1.0.6
lxml==6.0.2
MarkupSafe==2.1.5
memory-maze @ git+https://github.com/Nikita-Dudorov/memory-maze.git@ee4d78ab757c369d11ca42ef4d2b9c5614380773
mpmath==1.3.0
mujoco==3.3.7
networkx==3.5
numpy==2.3.5
pillow==11.3.0
protobuf==6.33.1
PyOpenGL==3.1.10
pyparsing==3.2.5
requests==2.32.5
scipy==1.16.3
sympy==1.14.0
torch==2.9.1+cu126
torchvision==0.24.1+cu126
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==2.0.1
zipp==3.23.0
